feat: add Loggable mixin
This was experimental in YAMDBF, and was buggy due to a Typescript bug, but it looks like that has since been fixed because it is working great now. Provides another way to easily attach the Logger to a class for ease of use, though really just using the static logging methods from the Logger module itself is already less characters than `this.logger` but to each their own. Just trying to be inclusive.